Style-fixes. Decomposing mui.theme. Added cross-buttons to clean inputs in some modals. Clean styles. Added breakpoint for 1200px to fontSize 12px.
