+++ odd-platform-ui/src/components/Management/DataSourcesList/DataSourceItem/DataSourceItemStyles.ts
@@ -4,7 +4,7 @@ export const styles = (theme: Theme) =>
   createStyles({
     container: {
       width: 'calc(100% - 8px)',
-      padding: theme.spacing(2, 2, 1, 2),
+      padding: theme.spacing(2, 2, 2, 2),
       margin: '0 auto',
       alignItems: 'flex-start',
       '&:hover': {
@@ -23,6 +23,14 @@ export const styles = (theme: Theme) =>
       display: 'none',
     },
     pullingValue: { color: 'green' },
+    descriptionContainer: {
+      '& > *': {
+        marginBottom: theme.spacing(1),
+      },
+      '& > *:last-child': {
+        marginBottom: theme.spacing(0),
+      },
+    },
   });
 
 export type StylesType = WithStyles<typeof styles>;
