+++ odd-platform-ui/src/components/Management/LabelsList/LabelCreateForm/LabelCreateFormItem/LabelCreateFormItem.tsx
@@ -1,18 +1,21 @@
 import React from 'react';
-import { withStyles, TextField } from '@material-ui/core';
+import { withStyles, TextField, InputAdornment } from '@material-ui/core';
 import { useFormContext, Controller } from 'react-hook-form';
 import AppButton from 'components/shared/AppButton/AppButton';
+import CancelIcon from 'components/shared/Icons/CancelIcon';
 import { styles, StylesType } from './LabelCreateFormItemStyles';
 
 interface LabelCreateFormItemProps extends StylesType {
   itemIndex: number;
   onItemRemove: () => void;
+  fieldsLength?: number;
 }
 
 const LabelCreateFormItem: React.FC<LabelCreateFormItemProps> = ({
   classes,
   itemIndex,
   onItemRemove,
+  fieldsLength,
 }) => {
   const { control } = useFormContext();
 
@@ -22,6 +25,7 @@ const LabelCreateFormItem: React.FC<LabelCreateFormItemProps> = ({
         name={`labels.${itemIndex}.name`}
         control={control}
         defaultValue=""
+        rules={{ required: true, validate: value => !!value.trim() }}
         render={({ field }) => (
           <TextField
             {...field}
@@ -29,13 +33,27 @@ const LabelCreateFormItem: React.FC<LabelCreateFormItemProps> = ({
             placeholder="Label Name"
             variant="outlined"
             name={`labels.${itemIndex}.name`}
+            InputProps={{
+              endAdornment: field.value && (
+                <InputAdornment position="start">
+                  <AppButton
+                    size="small"
+                    color="unfilled"
+                    icon={<CancelIcon />}
+                    onClick={() => field.onChange('')}
+                  />
+                </InputAdornment>
+              ),
+            }}
           />
         )}
       />
       <div className={classes.labelItemButtons}>
-        <AppButton size="small" color="dropdown" onClick={onItemRemove}>
-          Delete
-        </AppButton>
+        {fieldsLength && fieldsLength > 1 && (
+          <AppButton size="small" color="dropdown" onClick={onItemRemove}>
+            Delete
+          </AppButton>
+        )}
       </div>
     </>
   );
