+++ odd-platform-ui/src/components/Management/TagsList/TagsListStyles.ts
@@ -30,6 +30,7 @@ export const styles = (theme: Theme) =>
     },
     rowName: { color: '#A8B0BD' },
     listContainer: {
+      width: '100%',
       minHeight: '300px',
     },
     tagsItem: {
