+++ odd-platform-ui/src/theme/shadows.ts
@@ -1,0 +1,8 @@
+import { Shadows } from '@material-ui/core/styles/shadows';
+
+export const shadows: Shadows = new Array(25).fill('none') as Shadows;
+shadows[1] = '0 2px 6px 0 rgba(70, 73, 77, 0.2)';
+shadows[2] = '0 -2px 2px 0 rgba(0, 0, 0, 0.1)';
+shadows[3] = '0 1px 6px 0 rgba(70, 73, 77, 0.1)';
+shadows[8] = '0px 2px 6px rgb(0 0 0 / 10%)';
+shadows[9] = '0px 3px 12px rgba(0, 0, 0, 0.12)';
